+++ packages/inquiry/browser/index.js
@@ -1,6 +1,6 @@
-import RMIForm from './rmi-form.vue';
+import DefaultForm from './default-form.vue';
 
 export default (Browser, { component, mountPoint } = {}) => {
-  const InquiryForm = component || RMIForm;
+  const InquiryForm = component || DefaultForm;
   Browser.registerComponent('InquiryForm', InquiryForm, { mountPoint });
 };
